Rewrite command-line interface
Use helmsman to automatically find the commands, which are now
standalone scripts, and generate a list of commands.

Use commander.js to parse the command lines and generate per-command
help. Owing to a number of bugs and limitations, we use a patched
version, sigh. Upstream has been contacted about fixing things.

Remove ArgumentParser.