Import and use class names for points. Import point names into skeleton data.
Names in the name attribute are treated as whitespace separated.
If the left and right point share a name it is imported as is.
If the left point has a name that the right point has not it is prefixes by 'left-'
If the right point has a name that the left point has not it is prefixed by 'right-'

Thus if left has "hello world" and right has "hello univers" the result is
"hello left-world right-univers"

some language

repair setting class-names when loading a skeleton PenStrokePoint

remove console.log