fontinfo.plist import / export improvements.
The fontinfo.plist file is now imported when a master is imported.
When doing an export, the master's fontinfo.plist is read an becomes
part of the exported data.

In the future performing some cherry picking of the fontinfo.plist
from the masters to form part of the interpolated UFO instances would
be good.

Metadata like copyright / trademark is likely to be very similar if not
identical for masters in the same project.

On the other hand it may be the case that descender, ascender, and
italicAngle will need some treatment to be fully valid for exported
interpolated instances.