Interpolation command
Takes a list of masters and a list of proportions, and spits out
CPS (based on that written by Lasse for his initial demo) that
interpolates the given fonts with the given proportions.

To use the output, redirect it to a new .cps file under
data/com.metapolator/cps, and add a new master entry to project.yaml.

Or wait until I add this functionality myself.