Allow default port to be set for red-pill
Make red-pill default to port given by RED_PILL_PORT environment
variable. If  this is unset, the choice is random as before.

Fix a confusing error when an already-used port is used.

In the Vagrant box, forward the range of ports given by the environment
variable VAGRANT_RED_PILL_PORTS, set to a single value or single
inclusive range A-B; default 3000. Inside the Vagrant box, set
RED_PILL_PORT to the first port in VAGRANT_RED_PILL_PORTS.