ufo components: this can now render components again.
But this time it is using the data from the MOM model which includes
the component MOM object. Recursive component references are
supported, for example the reference from 'a' to 'b', in
components.ufo, where 'b' then references 'c'. The drawing of 'a' will
include {a,b,c}.

Transformations are not tested at all yet. So its rather likely that there
are some glitches in that section of the code.