Fix bug: RuleControler.prototype.reloadRules getting stuck after an error.
Before requesting new rules, we deleted the old rules
to clear the cache. This lead to a moment where requests
to `reloadRule` where answered with a KeyError, because
the rule was pruned. When reloading  of a rule failed, because
of a syntax bug in the cps for example, the reloadRule method
was blocked forever.