commander.js: merge yet more harmless code to get closer to upstream
As of now, we are about as close to upstream as we can get.

I am waiting for upstream to respond to my latest PR and will continue
to feed them a new one each time one is accepted until I have pushed all
our improvements and fixes upstream, or found upstream equivalents or
nice workarounds.