Fixed bug with range selection
Because the destacking removed the id value only when a parameter was
really stacked (multiple of the same operators), the range selection
(which depends on the id value) didn’t work properly.
Now the destacking is after each addOperator. This fixes the bug, but
is not optimal, because it is doing unnecessary destacking.

Cleaned up ‘self’ > ‘this’ as well on the way. These are redundant
since replacing angular for each with native for loops.