Fix hard to find cache related bug. Plus some better comments and cleaning up.
The bug was that AtNamespaceCollections did change the items of this._rules. In
the case of a child being another AtNamespaceCollection however, these items
were references to _rules cache of the child. Thus AtNamespace changed
the items of the child. After pruning the cache and reading the child items
again they also where manipulated again, now not matching any elements anymore,
because the namespaces was added twice/multiple times.
Now AtNamespaces makes a copy of every rule item that it changes.