Export metapolator apps for static deployments on a simple webserver
Used to be called 'export-static-red-pill

I included a help/usage message that reads:

Usage: build-static-deployment APPLICATION PROJECTDIR TARGETDIR

Build an instance of a Metapolator application as a HTML website
to be served (statically) from any simple HTTP file server.

APPLICATION: the Metapolator application to be exported, choices are:
	 * metapolator
	 * red-pill
	 * slider-demo
	 * stand-alone
	 * --help: displays this page
PROJECTDIR: a Metapolator project directory.
TARGETDIR: the target directory of this operation.