Big refactoring: select rules that apply to elements.
This was inspired by https://github.com/servo/servo/wiki/Css-selector-matching-meeting-2013-07-19

It makes the engine on initial execution just bit faster, (from > 7 seconds
selector-engine-time to <= 6 seconds on a heavy job of 30 seconds).
I was hoping for bigger gains. However, It shifts a lot of responsibility
into SelectorEngine and relieves ParamaterCollection and other CPS/{elements}
which is a rather good thing.