Update cloning/creation of UI-Masters and UI-Instances, use element properties more often.
In MOM UI-masters are prefixed with "master-" and UI-Instances are prefixed with "instance-".

Skeletons of instances are taken from their first master (but not yet updated when that master changes).

UI-Instances use for their state now just the element.properties of their MOM-Masters.

UI-Masters still use their cps files to store state, because that is more complex to switch.
The mechanics of the parameter require some more action.