Add simple check 'isInterpolationCompatible' to MOM-Master and its child Nodes.
Simple means that we don't allow concepts like 'patching' masters
here. Optionally it can be tested if one master would interpolate
with another master because it's a superset—the superfluous glyphs
are then omited. But within at the moment metapolator allowing
this would cause truble when the designspace is reordered i.e. the
interpolation base master changes.